+++ src/models/products/products.service.ts
@@ -4,7 +4,8 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Product, ProductDocument } from './entities/product.entity';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
-// import { Ingredient, IngredientDocument } from 'src/ingredients/entities/ingredient.entity';
+
+import mongoose = require('mongoose');
 
 @Injectable()
 export class ProductsService {
@@ -15,8 +16,8 @@ export class ProductsService {
 
   async create(createProductDto: CreateProductDto): Promise<ProductDocument> {
     try {
-      const productExist = this.findByName(createProductDto.name);
-      if (productExist) throw new ConflictException(`Procuct: ${createProductDto.name} already exists!`);
+      const hasProduct: ProductDocument = await this.productModel.findOne({ name: createProductDto.name })
+      if (hasProduct) throw new BadRequestException(`Procuct: ${createProductDto.name} already exists!`);
       const product = new this.productModel(createProductDto);
       // const hasIngredient = await this.ingredientModel.findOne(createProductDto.ingredients)
       await product.populate('ingredients').execPopulate();
@@ -30,11 +31,11 @@ export class ProductsService {
 
   async findAll(): Promise<ProductDocument[]> {
     try {
-      const products = await this.productModel.find().populate('ingredients');
+      const products: ProductDocument[] = await this.productModel.find().populate('ingredients');
       if (!(products.length)) throw new NotFoundException('No Products found.');
       products.forEach(product => {
         product.ingredients.forEach(ingredient => {
-          product.value += ingredient.value //not working?
+          product.cost += ingredient.value
         })
       });
       return products;
@@ -47,7 +48,8 @@ export class ProductsService {
 
   async findOne(id: string): Promise<ProductDocument> {
     try {
-      const product = await this.productModel.findById(id);
+      if (!this.validateObjectId(id)) throw new BadRequestException(`Invalid ObjectId sent!`);
+      const product: ProductDocument = await this.productModel.findById(id);
       if (!product) throw new NotFoundException('Product not found!');
       return product
     } catch (error) {
@@ -58,7 +60,7 @@ export class ProductsService {
 
   async findByName(name: string): Promise<ProductDocument> {
     try {
-      const product = await this.productModel.findOne({ name }).exec();
+      const product: ProductDocument = await this.productModel.findOne({ name }).exec();
       if (!product) throw new NotFoundException(`Procuct: ${name} was not found! `)
       return product;
     } catch (error) {
@@ -69,7 +71,8 @@ export class ProductsService {
 
   async updateOne(id: string, updateProductDto: UpdateProductDto): Promise<ProductDocument> {
     try {
-      const hasProduct = await this.productModel.findById(id);
+      if (!this.validateObjectId(id)) throw new BadRequestException(`Invalid ObjectId sent!`);
+      const hasProduct: ProductDocument = await this.productModel.findById(id);
       if (!hasProduct) throw new NotFoundException(`Prodfuct not found to update!`);
       return await this.productModel.findByIdAndUpdate(
         { _id: id },
@@ -84,7 +87,8 @@ export class ProductsService {
 
   async deleteOne(id: string): Promise<any> {
     try {
-      const product = await this.productModel.findById(id);
+      if (!this.validateObjectId(id)) throw new BadRequestException(`Invalid ObjectId sent!`);
+      const product: ProductDocument = await this.productModel.findById(id);
       if (!product) throw new NotFoundException(`Procuct not found to be deleted! `)
       return await this.productModel.deleteOne({ _id: id }).exec();
     } catch (error) {
@@ -93,18 +97,25 @@ export class ProductsService {
   }
 
   async productAvailable(id: string): Promise<any> {
-    var productAvailable = false;
-    const product = await this.productModel.findById(id).populate('ingredients').exec();
-    product.ingredients.forEach(ingredient => {
-      ingredient.quantity > 0 ? productAvailable = true : productAvailable = false;
-    });
+    try {
+      if (!this.validateObjectId(id)) throw new BadRequestException(`Invalid ObjectId sent!`);
+      const product: ProductDocument = await this.productModel.findById(id).populate('ingredients').exec();
+      if (!product) throw new NotFoundException(`Product not found!`);
+      var productAvailable: Boolean = false;
+      product.ingredients.forEach(ingredient => {
+        ingredient.quantity > 0 ? productAvailable = true : productAvailable = false;
+      });
 
-    return { response: { product: product.name, available: productAvailable } };
+      return { response: { product: product.name, available: productAvailable } };
+    } catch (error) {
+      throw error;
+    }
   }
 
   async uploadProductPicture(id: string, file: Express.Multer.File): Promise<ProductDocument> {
     try {
-      const product = await this.productModel.findById(id);
+      if (!this.validateObjectId(id)) throw new BadRequestException(`Invalid ObjectId sent!`);
+      const product: ProductDocument = await this.productModel.findById(id);
       if (!product) throw new NotFoundException('Upload failed, product not found!');
 
       //Image type validation
@@ -122,4 +133,28 @@ export class ProductsService {
     }
   }
 
+
+  async getCostReport(): Promise<any> {
+    try {
+      const products = await this.findAll();
+      let totalProductsCost = 0;
+      if (!products) throw new NotFoundException("No Products found!");
+      products.forEach((product) => {
+        product.ingredients.forEach(ingredient => {
+          totalProductsCost += ingredient.value;
+        })
+
+      })
+      return {
+        totalProductsCost: `The cost of all the products is: ${totalProductsCost}`,
+        products: products,
+      };
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  validateObjectId(id: string) {
+    return mongoose.Types.ObjectId.isValid(id);
+  }
 }
